Revive test_require.js
Details:
 * Add this test to the testrunner as `test_module_require.js`.
 * Move module `process` related code into `test_process_readsource.js`.
 ** Move the used JSON-file to `test/resources/process/`.

IoT.js-DCO-1.0-Signed-off-by: Zsolt Borbély zsborbely.u-szeged@partner.samsung.com
