+++ test/resources/process/package.json
