+++ test/run_pass/require1/require_add.js
@@ -17,8 +17,6 @@
   @NOTEST
 */
 
-var x = require("buffer");
-
 exports.add = function(a, b) {
   return a + b;
 };
