+++ test/run_pass/require1/test_index/add2.js
