+++ test/run_pass/require1/test_index/lib/multi.js
