+++ test/run_pass/require1/test_index/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "index_test"
+}
