+++ test/run_pass/require1/test_pkg/add2.js
