+++ test/run_pass/require1/test_pkg/lib/multi.js
