+++ test/run_pass/require1/test_pkg/main.js
