+++ test/run_pass/test_process_readsource.js
@@ -1,4 +1,4 @@
-/* Copyright 2015-present Samsung Electronics Co., Ltd. and other contributors
+/* Copyright 2017-present Samsung Electronics Co., Ltd. and other contributors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,8 +15,13 @@
 
 var assert = require('assert');
 
-var module_cache = require('run_pass/require1/module_cache.js');
-module_cache.i = 100;
-module_cache = require('run_pass/require1/module_cache.js');
+var json_file = process.cwd() + "/resources/process/package.json";
 
-assert.equal(module_cache.i, 100);
+// Load a JSON file.
+var str = process.readSource(json_file);
+var json = JSON.parse(str);
+
+assert.equal(json.version, "2.9.1");
+assert.equal(json.name, "npm");
+assert.equal(json.main, "./lib/npm.js");
+assert.equal(json.repository.type, "git");
