+++ test/testsets.json
@@ -47,6 +47,7 @@
     { "name": "test_i2c.js", "skip": ["all"], "reason": "need to setup test environment" },
     { "name": "test_iotjs_promise.js", "skip": ["all"], "reason": "es2015 is off by default" },
     { "name": "test_module_cache.js", "skip": ["nuttx"], "reason": "not implemented for nuttx" },
+    { "name": "test_module_require.js", "skip": ["nuttx"], "reason": "not implemented for nuttx" },
     { "name": "test_net_1.js" },
     { "name": "test_net_2.js" },
     { "name": "test_net_3.js", "timeout": 100, "skip": ["nuttx"], "reason": "requires too many socket descriptors and too large buffers" },
@@ -76,6 +77,7 @@
     { "name": "test_process_experimental_off.js", "skip": ["experimental"], "reason": "needed if testing stablity is set with stable" },
     { "name": "test_process_experimental_on.js", "skip": ["stable"], "reason": "needed if testing stablity is set with experimental" },
     { "name": "test_process_next_tick.js" },
+    { "name": "test_process_readsource.js" },
     { "name": "test_process_uncaught_order.js", "uncaught": true },
     { "name": "test_process_uncaught_simple.js", "uncaught": true },
     { "name": "test_pwm.js", "skip": ["all"], "reason": "need to setup test environment" },
